Change memory class to accept 32-bit address intead of 64-bit, also add getter for base address in core

+++ src/cortex-m0p_memory.cpp
@@ -1,32 +1,32 @@
 #include "cortex-m0p_memory.h"
 #include <cstdlib>
 
-BYTE CM0P_Memory:: read_byte(uint64_t address) {
+BYTE CM0P_Memory:: read_byte(uint32_t address) {
 	if (address >= size)
 		return 0;
 	return memory[address];
 }
 
-HALFWORD CM0P_Memory:: read_halfword(uint64_t address) {
+HALFWORD CM0P_Memory:: read_halfword(uint32_t address) {
 	return read_byte(address) << 8 | read_byte(address+1);
 }
 
-WORD CM0P_Memory:: read_word(uint64_t address) {
+WORD CM0P_Memory:: read_word(uint32_t address) {
 	return read_halfword(address) << 16 | read_halfword(address+2);
 }
 
 
-void CM0P_Memory:: write_byte(uint64_t address, BYTE data) {
+void CM0P_Memory:: write_byte(uint32_t address, BYTE data) {
 	if (address < size)
 		memory[address] = data;
 }
 
-void CM0P_Memory:: write_halfword(uint64_t address, HALFWORD data) {
+void CM0P_Memory:: write_halfword(uint32_t address, HALFWORD data) {
 	write_byte(address, data >> 8);
 	write_byte(address+1, data & 0xFF);
 }
 
-void CM0P_Memory:: write_word(uint64_t address, WORD data) {
+void CM0P_Memory:: write_word(uint32_t address, WORD data) {
 	write_halfword(address, data >> 16);
 	write_halfword(address+2, data & 0xFFFF);
 }
